refactor(engine): optimizasyon motoru Application katmanına taşındı (ALG-02)
Motor saf deterministik hesaplama; EF/DbContext/logger/HTTP kullanmıyor, bu
yüzden Infrastructure değil Application'a ait. ContaminationFilter,
OptimizationInput ve IOptimizationEngine zaten Application'daydı.

- Services/OptimizationEngine.cs -> Common/Optimization/OptimizationEngine.cs
- PlacedBox/UnplacedBox ayrı dosyaya çıkarıldı (private -> internal record)
- DI kaydı Application.AddApplication() içine alındı
- Engine.Tests artık yalnızca Application'a referans veriyor; EF/Hangfire/MinIO
  bağımlılıkları test çıktısından tamamen düştü
- InternalsVisibleTo Infrastructure'dan Application'a taşındı

Davranış sıfır değişim: golden-master snapshot'larının hiçbiri değişmedi.
Doğrulama: Passed 20 (Engine.Tests) + 20 (Infrastructure.Tests).
