+++ apps/backend/CargoPilot.Application/AssemblyInfo.cs
