+++ apps/backend/CargoPilot.Application/Common/Optimization/OptimizationEngine.cs
@@ -2,7 +2,7 @@ using CargoPilot.Application.Common.Interfaces;
 using CargoPilot.Application.Common.Models;
 using CargoPilot.Domain.Enums;
 
-namespace CargoPilot.Infrastructure.Services;
+namespace CargoPilot.Application.Common.Optimization;
 
 internal sealed class OptimizationEngine : IOptimizationEngine
 {
@@ -584,17 +584,4 @@ internal sealed class OptimizationEngine : IOptimizationEngine
         }
         return false;
     }
-
-    private sealed record PlacedBox(
-        Guid ItemId,
-        decimal X, decimal Y, decimal Z,
-        decimal W, decimal H, decimal D,
-        LoadingPlanPlacementRotation Rotation,
-        decimal Weight,
-        bool IsStackable,
-        int MaxStackCount,
-        decimal MaxWeightOnTop,
-        int? UnloadingOrder);
-
-    private sealed record UnplacedBox(Guid ItemId, UnplacedReason Reason);
 }
