+++ apps/backend/CargoPilot.Application/Common/Optimization/PlacedBox.cs
@@ -1,0 +1,16 @@
+using CargoPilot.Domain.Enums;
+
+namespace CargoPilot.Application.Common.Optimization;
+
+internal sealed record PlacedBox(
+    Guid ItemId,
+    decimal X, decimal Y, decimal Z,
+    decimal W, decimal H, decimal D,
+    LoadingPlanPlacementRotation Rotation,
+    decimal Weight,
+    bool IsStackable,
+    int MaxStackCount,
+    decimal MaxWeightOnTop,
+    int? UnloadingOrder);
+
+internal sealed record UnplacedBox(Guid ItemId, UnplacedReason Reason);
