+++ apps/backend/CargoPilot.Engine.Tests/CargoPilot.Engine.Tests.csproj
@@ -24,7 +24,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\CargoPilot.Infrastructure\CargoPilot.Infrastructure.csproj" />
+    <ProjectReference Include="..\CargoPilot.Application\CargoPilot.Application.csproj" />
   </ItemGroup>
 
 </Project>
