+++ apps/backend/CargoPilot.Infrastructure/DependencyInjection.cs
@@ -94,7 +94,6 @@ public static class DependencyInjection {
         services.AddScoped<IUserVehicleFavoriteRepository, UserVehicleFavoriteRepository>();
         services.AddScoped<ILoadingPlanRepository, LoadingPlanRepository>();
         services.AddScoped<ILoadingPlanItemGroupRepository, LoadingPlanItemGroupRepository>();
-        services.AddScoped<IOptimizationEngine, OptimizationEngine>();
         services.AddScoped<IIntegrationRepository, IntegrationRepository>();
         services.AddScoped<IPasswordResetTokenRepository, PasswordResetTokenRepository>();
         services.AddScoped<IUserPasswordHistoryRepository, UserPasswordHistoryRepository>();
